komorebi: stop setting windows to topmost
It is not komorebi's right or position to set this flag, it doesn't mean
the same thing as just a regular top of the z-order, it causes disorder
among child toolwindows and so on as well. This is also what was getting
in the way of the border window render.