komorebi: remove panic hook
It's unsafe to continue after a panic - many panic signify a condition
that can not be recovered from where memory may be in an invalid state.

This also blocks getting backtraces in debug mode.

This could be restored with something that terminates the process later,
after logging - but we could also redo the logging infrastructure to
redirect stderr to a file so as to have the standard handler handle
this, which is much easier/safer, as panic hooks are very hard to
implement correctly.